Make diff saner by not renaming `LemonDashboardHeader.tsx` yet
